refactor(web_search): move search + scrape tools into the env folder
Make web_search self-contained: relocate the top-level WebSearchToolkit
(tools/web_search.py) and WebScraperToolkit (tools/web_scraper.py) into
environments/web_search/ as search.py and scrape.py, imported relatively.
Kept as two files since the modules define a colliding DEFAULT_TIMEOUT.
